BlogController初步完善

+++ src/main/java/com/is/controller/BlogController.java
@@ -1,16 +1,25 @@
 package com.is.controller;
 
+import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.is.common.lang.Result;
 import com.is.entity.Blog;
+import com.is.entity.User;
 import com.is.service.IBlogService;
+import com.is.service.IUserService;
+import org.apache.shiro.SecurityUtils;
+import org.apache.shiro.authc.AuthenticationToken;
+import org.apache.shiro.authc.UsernamePasswordToken;
 import org.apache.shiro.authz.annotation.RequiresAuthentication;
+import org.apache.shiro.subject.Subject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Objects;
 
 /**
  * <p>
@@ -26,6 +35,8 @@ import java.util.List;
 public class BlogController {
     @Autowired
     IBlogService blogService;
+    @Autowired
+    IUserService userService;
 
     @GetMapping("/{id}")
     public Result detail(@PathVariable Integer id) {
@@ -38,16 +49,31 @@ public class BlogController {
 
     @PostMapping
     public Result add(@RequestBody Blog blog) {
+        Subject subject = SecurityUtils.getSubject();
+        User user = (User) subject.getPrincipal();
+        blog.setUserId(user.getId());
+        blog.setStatus(true);
+        blog.setCreated(LocalDateTime.now());
         boolean flag = blogService.save(blog);
-        return flag?Result.succ("发布成功"):Result.fail("发布失败");
+        return flag?Result.succ("发布成功",blog):Result.fail("发布失败");
     }
     @PutMapping
     public Result update(@RequestBody Blog blog) {
+        Subject subject = SecurityUtils.getSubject();
+        User user = (User) subject.getPrincipal();
+        if(!user.getId().equals( blog.getUserId())&& !user.getPerms().equals("admin") ){
+            return Result.fail("没有修改权限");
+        }
         boolean flag = blogService.updateById(blog);
-        return flag?Result.succ("修改成功"):Result.fail("修改失败");
+        return flag?Result.succ("修改成功",blog):Result.fail("修改失败");
     }
     @DeleteMapping("/{id}")
     public Result delete(@PathVariable Integer id) {
+        Subject subject = SecurityUtils.getSubject();
+        User user = (User) subject.getPrincipal();
+        if(!user.getId().equals(blogService.getById(id).getUserId())&& !user.getPerms().equals("admin") ){
+            return Result.fail("没有删除权限");
+        }
         boolean flag = blogService.removeById(id);
         return flag?Result.succ("删除成功"):Result.fail("删除失败");
     }
@@ -58,11 +84,16 @@ public class BlogController {
 
     }
     @GetMapping("/{currentPage}/{size}")
-    public Result blogs(@PathVariable Integer currentPage) {
-        if(currentPage == null || currentPage < 1) currentPage = 1;
-        Page page = new Page(currentPage, 5);
-        IPage pageData = blogService.page(page, new QueryWrapper<Blog>().orderByDesc("created"));
-        return Result.succ(pageData);
+    public Result blogPages(@PathVariable Integer currentPage, @PathVariable Integer size) {
+        LambdaQueryWrapper<Blog> blogLambdaQueryWrapper = new LambdaQueryWrapper<>();
+        blogLambdaQueryWrapper.isNotNull(Blog::getId);
+        Page<Blog> page = blogService.page(new Page<>(currentPage, size), blogLambdaQueryWrapper);
+        if( currentPage > page.getPages()) {
+            page = blogService.page(new Page<>(page.getPages(), size), blogLambdaQueryWrapper);
+        }
+
+
+        return Result.succ(page);
     }
 
 }
